Starting to use genieutils instead of geniedat. Loading/saving not done yet.

+++ AGE_ComboBoxes.cpp
@@ -161,7 +161,7 @@ void ComboBox_Long::OnKillFocus(wxFocusEvent& Event)
 	((TextCtrl_Long*)Event.GetEventObject())->OnKillFocus(Event);
 	if(!((TextCtrl_Long*)Event.GetEventObject())->NoLoadList)
 	{
-		long Value = lexical_cast<long>(((TextCtrl_Long*)Event.GetEventObject())->GetValue());
+		int Value = lexical_cast<int>(((TextCtrl_Long*)Event.GetEventObject())->GetValue());
 		((ComboBox_Long*)((TextCtrl_Long*)Event.GetEventObject())->ParentContainer)->SetSelection(0);
 		((ComboBox_Long*)((TextCtrl_Long*)Event.GetEventObject())->ParentContainer)->SetSelection(Value + 1);
 	}
